Demote consequence of not having C++11 support
This module sets individual flags for each feature; there is not one
"CXX11Features_FOUND" flags which is set. Thus, when putting the
module as REQUIRED, the only consequence is that we cannot disable
searching for it.

Instead, we make it optional, meaning that the flags won't get set if
it isn't included and the dependent modules will adjust accordingly.